# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Api
 * AEO Chatbot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
  messageCount: number;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
  /** @nullable */
  tokensUsed: number | null;
  /** @nullable */
  responseTimeMs: number | null;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messageCount: number;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface DashboardStats {
  totalConversations: number;
  totalMessages: number;
  /** @nullable */
  avgResponseTimeMs: number | null;
  totalTokensUsed: number;
  conversationsToday: number;
  messagesLast7Days: number;
}

export interface ActivityItem {
  id: number;
  conversationId: number;
  conversationTitle: string;
  lastMessage: string;
  role: string;
  createdAt: string;
}

export interface DailyVolume {
  date: string;
  messageCount: number;
  conversationCount: number;
}

export interface BackendLog {
  id: number;
  timestamp: string;
  event: string;
  model: string;
  /** @nullable */
  conversationId: number | null;
  /** @nullable */
  tokensUsed: number | null;
  /** @nullable */
  responseTimeMs: number | null;
  status: string;
  /** @nullable */
  details: string | null;
}

export interface KeywordItem {
  keyword: string;
  brand: string;
  locked: boolean;
  last_score: number;
  days_idle: number;
  consecutive_top: number;
  rotation_priority: number;
}

export interface KeywordCluster {
  cluster: string;
  keywords: KeywordItem[];
  locked_count: number;
  total_count: number;
}

export interface KeywordRotationStatus {
  clusters: KeywordCluster[];
  /** @nullable */
  last_updated: string | null;
}

export interface RotationTriggerInput {
  /** @nullable */
  cluster?: string | null;
}

export interface RotationTriggerResponse {
  status: string;
  cluster: string;
  message: string;
}

export interface InjectKeywordItem {
  keyword: string;
  ground_truth: string;
}

export type InjectClusterInputBacklinksItem = { [key: string]: unknown };

export interface InjectClusterInput {
  business_name: string;
  brand: string;
  keywords: InjectKeywordItem[];
  backlinks?: InjectClusterInputBacklinksItem[];
}

export interface InjectClusterResponse {
  cluster: string;
  keyword_count: number;
  message: string;
}