# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEO Chatbot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /keyword-rotation/inject-cluster:
    post:
      operationId: injectKeywordCluster
      tags: [keyword-rotation]
      summary: Inject business analysis keywords into the rotation system
      description: >
        Takes keywords produced by the AEO business analysis and writes them as
        a named cluster in keyword_clusters.yaml. Existing locked keywords are
        preserved. New keywords start at score 0.0 and are immediately queued
        for the next rotation cycle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InjectClusterInput"
      responses:
        "200":
          description: Cluster written successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InjectClusterResponse"
...
  /keyword-rotation/run:
    post:
      operationId: triggerKeywordRotation
      tags: [keyword-rotation]
      summary: Trigger a keyword rotation cycle
      description: >
        Starts a background keyword rotation run. Optionally scoped to a single
        cluster. Poll /keyword-rotation/status for updated scores after the cycle
        completes (typically 30–120 seconds).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RotationTriggerInput"
      responses:
        "200":
          description: Rotation job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotationTriggerResponse"
components:
  schemas:
...
    RotationTriggerInput:
      type: object
      properties:
        cluster:
          type: ["string", "null"]
      required: []
    RotationTriggerResponse:
      type: object
      properties:
        status:
          type: string
        cluster:
          type: string
        message:
          type: string
      required:
        - status
        - cluster
        - message
    InjectKeywordItem:
      type: object
      properties:
        keyword:
          type: string
        ground_truth:
          type: string
      required:
        - keyword
        - ground_truth
    InjectClusterInput:
      type: object
      properties:
        business_name:
          type: string
        brand:
          type: string
        keywords:
          type: array
          items:
            $ref: "#/components/schemas/InjectKeywordItem"
        backlinks:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - business_name
        - brand
        - keywords
    InjectClusterResponse:
      type: object
      properties:
        cluster:
          type: string
        keyword_count:
          type: integer
        message:
          type: string
      required:
        - cluster
        - keyword_count
        - message