# lib/api-zod/src/generated/types/injectClusterInput.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Api
 * AEO Chatbot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InjectClusterInputBacklinksItem } from './injectClusterInputBacklinksItem';
import type { InjectKeywordItem } from './injectKeywordItem';

export interface InjectClusterInput {
  business_name: string;
  brand: string;
  keywords: InjectKeywordItem[];
  backlinks?: InjectClusterInputBacklinksItem[];
}