# lib/api-zod/src/generated/types/injectKeywordItem.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Api
 * AEO Chatbot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InjectKeywordItem {
  keyword: string;
  ground_truth: string;
}